tests/wpt/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.perspective.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.transformation.perspective</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.perspective</h1>
<p class="desc">perspective() results in the correct transformation matrix</p>


<script>
var t = async_test("perspective() results in the correct transformation matrix");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

const length = 100;
ctx.perspective(length);
const domMatrix = new DOMMatrix();
domMatrix.m34 = -1/length;
_assertMatricesApproxEqual(domMatrix, ctx.getTransform());
ctx.rotateAxis(1, 2, 3, 4);
domMatrix.rotateAxisAngleSelf(1, 2, 3, rad2deg(4));
_assertMatricesApproxEqual(domMatrix, ctx.getTransform());
t.done();

});
</script>
